move index.css to css folder
